+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 * Add `aniFirstNameResponseCode` and `aniLastNameResponseCode` to `CreditCardVerification`
+* Add `shippingTaxAmount` to `Transaction` and `TransactionRequest`
 
 ## 3.35.0
 * Add attributes to `ThreeDSecureLookupRequest`